data-pipeline/crawler/requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.3
newspaper3k>=0.2.8
lxml>=5.3.0
certifi>=2024.7.4

